refactor(predictions): break the league-icon import cycle with a const leaf module
