+++ src/features/polymarket/components/league-icon/iconConstants.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_SIZE = 24;
+
+export type IconProps = { color: string; width?: number; height?: number };
